fix: rely on lazy tenant initialisation instead of bulk initialisation after startup
Bulk initialisation has caused issues in the past. Lazy initialisation seems to be the most viable strategy if we need to handle migrations with breaking changes. If we really want to optimise away the resultant latency for the first request to a tenant, we could add support to trigger a gradual initialisation after a rolling update is complete.