fix: inject a fresh metrics registry in tests
fastify-metrics/prom-client by default uses a global registry, which
results in multiple initializations of the app failing since they'll
try to register duplicate metrics in the registry.

This usage pattern only occurs in our test code, which has now been
updated to inject a fresh registry for each app initialization, in
order to sidestep this issue.