chore: remove reference to old comment
Previously, before the migration to `fastify-metrics` v9, the v8
configuration included a blacklist property which seemed like it
attempted to hide a few metrics. However, it appears that the
`blacklist` property wasn't intended for hiding metrics, and doesn't
seem like there's such a configuration option in v9 either.

Hence, this commit removes the leftover comment.