fix: allow connections via IP address as host (#641)
* fix: allow connections via IP address as host

Currently, we're passing in a root CA for all requests as SSL option. However, when passing in an IP address, this will always fail. As we are starting to transition storage over to pgbouncer, internal connections require connecting via an IPv6 address. To support this, we need to not include the root CA options, in case we're connecting against an IP.

* unify

* decode uri component